refactor project to use cobra cli framework
