+++ go.mod
@@ -1,3 +1,10 @@
 module github.com/matthew-c-atu/project-audio-streamer
 
 go 1.22.1
+
+require github.com/spf13/cobra v1.8.0
+
+require (
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+)
