+++ main.go
@@ -1,0 +1,11 @@
+/*
+Copyright © 2024 NAME HERE <EMAIL ADDRESS>
+
+*/
+package main
+
+import "github.com/matthew-c-atu/project-audio-streamer/cmd"
+
+func main() {
+	cmd.Execute()
+}
